+++ actions/pdf/pdfjs/web/magazine.js
@@ -47,7 +47,7 @@ var MagazineView = {
         }
 		if(window.location.hash.indexOf('background=') > -1) { // background color
 			ix = window.location.toString().indexOf('background=') + 11;
-			col = '#'+window.location.toString().substring(ix);
+			col = '#'+window.location.toString().substring(ix,ix+6);
 			$('body').css('background',col);
 			$('#toolbarContainer').css('background',col);
 			$('#toolbarContainer').css('background','url(../web/images/texture.png), linear-gradient('+col+', #5b5a5a)');
@@ -94,6 +94,14 @@ var MagazineView = {
 		}
 
         $("#magazine").show();
+        
+		if(window.location.hash.indexOf('bgbtns=') > -1) { // background buttons color
+			ix = window.location.toString().indexOf('bgbtns=') + 7;
+			col = '#'+window.location.toString().substring(ix,ix+6);
+			$('.icon-holder').css('background-color',col);
+			$('.icon-center-holder').css('background-color',col);
+            $('.flashy-close').css('background-color',col);
+		}
 
         var pages = [1];
 		MagazineView.layout =  isPhone ? 'single' : 'double';